+++ front-end/src/App.js
@@ -1,11 +1,10 @@
-import React, { Component } from 'react';
+import React, { Component } from 'react'
 // import './App.css';
-import {Switch, Route} from "react-router-dom";
-import Stripe from "./Component/Stripe/Stripe.js"
-import LoginForm from "./Component/loginform";
-import Register from "./Component/Register/Register";
-import LandingPage from "./Component/LandingPage/LandingPage";
-import Navbar from "./Component/Navbar/Navbar";
+import {Switch, Route} from 'react-router-dom'
+import Stripe from './components/stripe'
+import LoginForm from './components/loginform'
+import LandingPage from './components/landingpage.js'
+import Navbar from './components/navbar'
 
 class App extends Component {
   render() {
@@ -13,7 +12,6 @@ class App extends Component {
       <div>
         <Switch>
           <Route exact path="/login" component={LoginForm} />
-          <Route exact path="/register" component={Register} />
           <Route exact path="/" component={LandingPage} />
           <Route exact path="/stripe" component={Stripe}/>
           <Route exact path="/navbar" component={Navbar}/>
