--- front-end/src/Component/LandingPage/LandingPage.css
@@ -1,13 +1,0 @@
-.article {
-    height: 300px;
-    width: 90%;
-}
-
-.swiper-slide {
-    height: 40px;
-    width: 80px;
-}
-
-.article {
-    width: 80%;
-}
