+++ front-end/src/components/css/landing-page-about-us-card.css
