+++ front-end/src/components/css/landing-page-home-card.css
